+++ src/extension/features/budget/subtract-upcoming-from-available/budget-table-row.js
@@ -1,0 +1,48 @@
+import { formatCurrency } from 'toolkit/extension/utils/currency';
+import { getEmberView } from 'toolkit/extension/utils/ember';
+import { getSelectedMonth } from 'toolkit/extension/utils/ynab';
+import * as categories from 'toolkit/extension/features/budget/subtract-upcoming-from-available/categories';
+import * as util from 'toolkit/extension/features/budget/subtract-upcoming-from-available/util';
+
+export function handleBudgetTableRow(element) {
+  if (!util.shouldInvoke()) return;
+
+  const category = getEmberView(element.id, 'category');
+  if (category) updateCategoryAvailableBalance(category, element);
+}
+
+function updateCategoryAvailableBalance(category, context) {
+  const categoryData = categories.getCategoryData(
+    categories.getCategoriesObject(),
+    getSelectedMonth(),
+    category.categoryId
+  );
+  if (!categoryData) return;
+
+  const $available = $(`.ynab-new-budget-available-number`, context);
+  const $availableText = $(`.user-data`, $available);
+
+  $availableText.text(formatCurrency(categoryData.availableAfterUpcoming));
+
+  $available.children('svg.icon-upcoming').remove();
+
+  const classes = 'upcoming positive zero negative';
+  $available.removeClass(classes);
+  $availableText.removeClass(classes);
+
+  const currencyClass = util.getCurrencyClass(categoryData.availableAfterUpcoming);
+  $available.addClass(currencyClass);
+  $availableText.addClass(currencyClass);
+
+  if (categoryData.availableAfterUpcoming >= 0) {
+    $(category).removeAttr('data-toolkit-negative-available');
+
+    if (category.isOverSpent) {
+      $available.addClass('cautious');
+      $availableText.addClass('cautious');
+    }
+  } else if (!category.isOverSpent) {
+    $available.removeClass('cautious');
+    $availableText.removeClass('cautious');
+  }
+}
