+++ src/extension/features/budget/subtract-upcoming-from-available/util.js
@@ -1,0 +1,37 @@
+import { formatCurrency } from 'toolkit/extension/utils/currency';
+import { l10n } from 'toolkit/extension/utils/toolkit';
+import { isCurrentRouteBudgetPage } from 'toolkit/extension/utils/ynab';
+
+export function shouldInvoke() {
+  return isCurrentRouteBudgetPage();
+}
+
+export function getCurrencyClass(amount) {
+  let currencyClass = 'positive';
+
+  if (amount < 0) {
+    currencyClass = 'negative';
+  } else if (amount === 0) {
+    currencyClass = 'zero';
+  }
+
+  return currencyClass;
+}
+
+export function createBudgetBreakdownEntry(elementId, l10nKey, l10nDefault, amount) {
+  const title = l10n(l10nKey, l10nDefault);
+
+  const currencyClass = getCurrencyClass(amount);
+  amount = formatCurrency(amount);
+
+  return $(`
+      <div id="${elementId}">
+        <div>${title}</div>
+        <div class="user-data">
+          <span class="user-data currency ${currencyClass}">
+            ${amount}
+          </span>
+        </div>
+      </div>
+    `);
+}
